test_container: Add non-layered demo test image
Often our tests are based on kuryr/demo. That image used to be based on
python and alpine and was taking a very long time to pull for already
slow tests. This change makes this repository have the authoritative
version of the kuryr/demo build.

The way it is built is just by running mkrootfs.sh. But that only needs
to be run if we change the version of curl or modify server.go.
Otherwise, to run test it just needs:

    docker build -t kuryr/demo . -f Dockerfile

This build will not need to go to Docker Hub for anything so we also
eliminate one source of flakiness.

Depends-On: https://review.openstack.org/#/c/547390/
Change-Id: Iab5799e22c3b353a4f2adbaa280ab4a9904ecd95
Signed-off-by: Antoni Segura Puimedon <antonisp@celebdor.com>