Add tempest coverage for namespace creation
It adds coverage to test namespace creation when the namespace
handler and subnet driver are used. It checks the namespace gets
created but also that the related resources get created too, in this
case the Neutron Subnet, as well as the Kuryr Net CRD. In addition,
it check the KuryrNet CRD has the right information about the network
and subnet IDs created.

Depends-On: I84580201f38c219f1943510bb493da0f07e07153

Change-Id: Iafc08ede300aecf1dc52135c6e51b89875e729d6