Increase ping counter to avoid tests flakiness
There may be cases where sending just one packet is not enough or
may randomly fail. One example for this is with OVN backend where
the first packet will be drop by design if it needs to go through
a logical router. For this reason this patch increases the ping
count to 4 to be on the safe side.

Change-Id: Ic2bdc3f75c684cc4a9c99b10268cf9390e3e4586