Explicitly disable service tests by default
This commit sets a new config option that will disable service tests
before they go to the 'resource_setup' phase, as just adding a skip
decorator to the tests still triggers the setup phase and breaks the
gate.

As it happened with [1] this commit will be reverted as soon as we
identify the root cause for the issue.

[1] https://review.openstack.org/#/c/577821/

Change-Id: I2452bde17c56e94d2942e5a293e2e4ccc3e1b2b3
Partial-Bug: #1778483