Get rid of cmp() usage and randomize pod labels
cmp() builtin is not available in Python 3. This commit removes its
usages. It also turns out we were comparing int to string in
assertNotEqual() in Services tests, so that was always passing.

Also tests were interfering with each other as labels added for pods in
Services tests were hardcoded. That's fixed here as well.

Change-Id: I37a5805f7670f905db24ddf86d3a4b4839453a6d