Set IfNotPresent imagePullPolicy for created pods
Even though kuryr-tempest-plugin DevStack plugin is building the
kuryr/demo container to make it available locally, it seems that we
create pods with the default imagePullPolicy, which is Always. This
means that image is downloaded anyway.

To work this around we can just set imagePullPolicy to IfNotPresent and
this commit does so.

Please note that an alternative could be to use a specified image tag
instead of :latest, but IfNotPresent still seems to make sense in tests.

Change-Id: I1fa559da39a44d802edb6f46d71cb8d20b66fd7a
Closes-Bug: 1799239