Basic IPv6 support
This commit fixes stuff related to the fact that most apps require IPv6
addresses to be in brackets. Also now tests that use floating ips are
skipped if IPv6 is configured. FIPs are not supported on IPv6 and we
need to figure out other way to test them.

Implements: blueprint kuryr-ipv6-support
Change-Id: Ic2be3cf93bd9d114af907d26198e8011281bfabf