Fixing test_ipblock_network_policy_sg_rules
The test was unstable, because it was only waiting for the SG to get
created assuming rules are created in the same moment. It's not true,
rules will follow and we need to account for that.

Commit fixes that by making sure we're waiting until all the rules are
in the status of KuryrNetworkPolicy.

Closes-Bug: 1860554
Depends-On: I3bbac8e90553be9424fa0ecdb4c05e096da8f2d7
Change-Id: I7fb66c86982b4b7d3e3296cbecda0fd6839b1177