Increase timeout to read pod commands results
exec_command_in_pod was hanging from time to time and we've set a 10
seconds timeout waiting for a response from it. This might be too little
as we very often see empty responses being returned. This commit
attempts to solve the issue by setting default timeout to 60 seconds.
Also the method used is changed from connect_get_namespaced_pod_exec to
connect_post_namespaced_pod_exec. This is what `kubectl` uses to
perform `exec`, so hopefully it'll be more reliable.

Depends-On: I3bbac8e90553be9424fa0ecdb4c05e096da8f2d7
Change-Id: I49ac4d066a0bd0da3833dcbf41655f7c9e374fd9