Ensure lb is ACTIVE before connectivity check
When the load balancer reconciliation is triggered
besides checking if the current lb ID is different
from the previous lb, the lb members should also be
configured. This commit enforces the existence of the
lb members. Also, it checks if the load balancer is
ACTIVE before checking connectivity.

Note, that it also removes the tempests gates to avoid
a circular dependency.

Change-Id: I64d8596f8d0504eeb898d40be7a11aa6635bca17