Change get_readiness_state() return type
Remove the usage of the bool() method as it wasn't behaving as expected
- bool("False") returns True instead of False.

The get_readiness_state() method was returning a string when
container_name wasn't None and it's been changed to return a boolean
instead.
container.ready param is already a Boolean but condition.status is a
string.

Increase the timeout waiting for pods to be ready after scaling in the
scale_deployment() method.

Change-Id: Ie64f868b877ca4c9f1ff8916360638f013fc2926