Add wait until active LB in NP test
The test test_egress_np_to_service_wo_selector is failing in some
environments when checking the connectivity to services (before
the NP is applied). The endpoints are defined in kuryrloadbalancers
objects but the connectivity is not working.

This change adds the method wait_until_service_LB_is_active() and
calls it in that test before checking the connectivity to the
services.

Change-Id: I41cec1ea1db43e2fbeb0bf6d5fafb84c87e77d1c