Temporarily disable stable jobs
We have multiple CI stability issues going on at the moment.  In order
to be able to merge master fixes, we need to disable broken stable jobs.
Once kuryr-tempest-plugin is stabilized we should be able to merge fixes
to the stable jobs and then we'll be able to revert this patch and make
them voting again.

Change-Id: I5e8fccec80b1a5593e9e32f8dadf8bc220d0424a