Use one pod to check connectivity in kuryr_restart
In test_kuryr_restart use one pod to check connectivity to the
created pods from this pod instead of assigning a fip for each created
pod. This can solve problems when there is shortage of floating ips.

Also, Don't fail on deletion of a pod in cleanup if it's not found

Change-Id: I55a33f88356ff3d1ef6ed0ce83bb102172e46023